rename recoveryword model so that its a viewmodel.
